backend/TG-back/requirements.txt
amqp==5.2.0
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
cryptography==43.0.1
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
redis==5.1.0
requests==2.32.3
requests-file==2.1.0
six==1.16.0
tldextract==5.1.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
